Change for deserialization into BigDecimal.

+++ api/src/test/java/com/senorpez/trident/api/PlanetaryCalendarControllerTest.java
@@ -12,6 +12,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 import java.io.InputStream;
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
@@ -1513,7 +1514,7 @@ public class PlanetaryCalendarControllerTest {
                 .andExpect(jsonPath("$.caste", instanceOf(Integer.class)))
                 .andExpect(jsonPath("$.day", instanceOf(Integer.class)))
                 .andExpect(jsonPath("$.shift", instanceOf(Integer.class)))
-                .andExpect(jsonPath("$.tithe", instanceOf(Double.class)))
+                .andExpect(jsonPath("$.tithe", instanceOf(BigDecimal.class)))
                 .andExpect(jsonPath("$._links.index", hasEntry("href", "http://localhost:8080/")))
                 .andExpect(jsonPath("$._links.self", hasEntry("href", String.format(
                         "http://localhost:8080/systems/%d/stars/%d/planets/%d/calendars/%d/currentTime",
@@ -1572,7 +1573,7 @@ public class PlanetaryCalendarControllerTest {
                 .andExpect(jsonPath("$.caste", instanceOf(Integer.class)))
                 .andExpect(jsonPath("$.day", instanceOf(Integer.class)))
                 .andExpect(jsonPath("$.shift", instanceOf(Integer.class)))
-                .andExpect(jsonPath("$.tithe", instanceOf(Double.class)))
+                .andExpect(jsonPath("$.tithe", instanceOf(BigDecimal.class)))
                 .andExpect(jsonPath("$._links.index", hasEntry("href", "http://localhost:8080/")))
                 .andExpect(jsonPath("$._links.self", hasEntry("href", String.format(
                         "http://localhost:8080/systems/%d/stars/%d/planets/%d/calendars/%d/currentTime",
